# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "全台藥局口罩資訊",
        "description": "全台藥局口罩資訊 OpenApi description",
        "contact": {
            "email": "conscience1113@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "L5 Swagger OpenApi dynamic host server"
        },
        {
            "url": "https://projects.dev/api/v1",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
        "/api/pharmacies": {
            "get": {
                "tags": [
                    "Pharmacies"
                ],
                "summary": "Get pharmacies by addresses or name",
                "description": "Returns pharmacies data",
                "operationId": "getPharmaciesByAddressOrName",
                "parameters": [
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address you search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name you search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:pharmacies",
                            "read:pharmacies"
                        ]
                    }
                ]
            }
        },
        "/api/pharmacies/all": {
            "get": {
                "tags": [
                    "Pharmacies"
                ],
                "summary": "Get all pharmacies",
                "description": "Returns pharmacies data",
                "operationId": "getPharmacies",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:pharmacies",
                            "read:pharmacies"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
    "tags": [
        {
            "name": "Pharmacies",
            "description": "get pharmacies data",
            "externalDocs": {
                "description": "Find out more",
                "url": "http://swagger.io"
            }
        }
    ],
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}